+++ packages/edge-login-ui-rn/src/native/components/common/ListItems/TextRowComponent.js
@@ -20,10 +20,11 @@ type State = {
 type Props = OwnProps & DispatchProps
 class TextRowComponent extends Component<Props, State> {
   numberOfLines: number
-  componentWillMount () {
-    this.setState({
+  constructor (props: Props) {
+    super(props)
+    this.state = {
       pressed: false
-    })
+    }
     this.numberOfLines = this.props.numberOfLines || 1
   }
   _onPressButton = () => {