+++ packages/edge-login-ui-rn/src/native/components/screens/existingAccout/ChangeAccountPinScreenComponent.js
@@ -28,13 +28,14 @@ export default class ChangeAccountPinScreenComponent extends Component<
   Props,
   State
 > {
-  componentWillMount () {
-    this.setState({
+  constructor (props: Props) {
+    super(props)
+    this.state = {
       username: '',
       pin: '',
       isProcessing: false,
       focusOn: 'pin'
-    })
+    }
   }
   renderHeader = (style: Object) => {
     if (this.props.showHeader) {