+++ packages/edge-login-ui-rn/src/native/components/screens/newAccount/TermsAndConditionsScreenComponent.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { AbcAccount } from 'edge-core-js'
+import type { EdgeAccount } from 'edge-core-js'
 import React, { Component } from 'react'
 import { Text, View } from 'react-native'
 
@@ -12,9 +12,9 @@ import SafeAreaView from '../../common/SafeAreaViewGradient.js'
 
 type Props = {
   styles: Object,
-  accountObject: AbcAccount,
+  accountObject: EdgeAccount,
   terms: Object,
-  agreeToCondition(AbcAccount): void
+  agreeToCondition(EdgeAccount): void
 }
 type State = {
   totalChecks: number
@@ -23,10 +23,11 @@ export default class TermsAndConditionsScreenComponent extends Component<
   Props,
   State
 > {
-  componentWillMount () {
-    this.setState({
+  constructor (props: Props) {
+    super(props)
+    this.state = {
       totalChecks: 0
-    })
+    }
   }
   renderItems (style: Object) {
     return this.props.terms.items.map(Item => (