+++ packages/edge-login-ui-rn/src/native/styles/screens/RecoverPasswordSceneStyles.js
@@ -1,18 +1,15 @@
 // @flow
 
 import * as Constants from '../../../common/constants'
-import * as Styles from '../'
 import { scale } from '../../../common/util'
+import * as Styles from '../'
 
 const RecoverPasswordSceneStyles = {
   screen: { ...Styles.ScreenStyle },
   header: {
-    ...Styles.HeaderContainerStyle,
+    ...Styles.HeaderContainerScaledStyle,
     backgroundColor: Constants.PRIMARY
   },
-  /* gradient: {
-    height: THEME.HEADER
-  }, */
   body: {
     padding: scale(18)
   },
@@ -27,7 +24,6 @@ const RecoverPasswordSceneStyles = {
   buttonContainer: {
     width: '100%',
     alignItems: 'center'
-    // height: THEME.BUTTONS.HEIGHT
   },
   modalMiddle: {
     width: '100%'
@@ -47,7 +43,10 @@ const RecoverPasswordSceneStyles = {
   },
   inputModal: {
     ...Styles.MaterialInputOnWhite,
-    container: { ...Styles.MaterialInputOnWhite.container, width: '100%' }
+    container: {
+      position: 'relative',
+      width: '100%'
+    }
   },
   inputError: {
     ...Styles.MaterialInputOnWhite,
@@ -98,7 +97,7 @@ const RecoverPasswordSceneStyles = {
   staticModalText: {
     color: Constants.GRAY_1,
     width: '100%',
-    fontSize: scale(15),
+    fontSize: scale(13),
     textAlign: 'center'
   },
   listItem: Styles.ListItemTextOnly,